Refactor the recursive html enriching
